Update spectroscopic_degraders.py
Several changes: added two new configuration parameters so the user can input ratios based on any survey, but defaults to the HSC based selection. User inputs an hdf5 file with the ratios in each pixel, and a pickled dictionary specifying which band was used for the magnitude, and which bands were used to compute the colors, as well as the magnitude and color limits for whatever survey they want to use.
For pixels with galaxies and ratios that do not result in a whole number of galaxies in the training sample, implements a random selection to choose whether to keep the extra galaxy or not.